1. apple oauth 로그인에서만 사용하는 Key 객체를 Inner class로 이동
